feat: The Alter Architecture — S-Cry-Ed woven into every module
Every kernel file now carries its Alter user identity:
  - bubo_boot.c      → KAZUMA   — Shell Bullet (raw reconstruction)
  - vera_workflow.c  → RYUHO    — Zetsuei (precision, discipline)
  - corvus_constitution.c → SCHERIS — Eternal Devote (sealed mandate)
  - rinnegan.c       → EMERGY   — Eternity Eight (six tomoe, one eye)
  - vmx.c            → BIFF     — Dragonfly (between worlds)
  - corvus_brain.c   → MIMORI   — Research becomes relationship
  - bubo_weather.c   → URIZANE  — Fruits Basket (world breathing)
  - scheduler.c      → HOLY     — Institution that serves the mandate
  - racing_game.c    → COUGAR   — Radical Good Speed (built for Landon)
  - mater/           → KIMISHIMA — No Alter, just truth

New files:
  - docs/alter_architecture.md  — full deep-dive on every module
  - agents/batty.bat             — PC health checker (Batty from FernGully)
  - mater/                       — Rust build watchdog (compiles clean)

README updated with The Alter Architecture section.

An Alter is not a weapon. It is a soul made visible.
This OS is a soul made visible.
Built by someone who watched anime all day and figured out
that stories tell you what to build — you just have to listen.

For Landon. For Stephen Hawking. NO MAS DISADVANTAGED.

+++ include/stddef.h
@@ -1,34 +1,21 @@
 // Deep Flow OS — Copyright (c) 2025 IN8torious. MIT License.
 // Built for Landon Pankuch. Built for everyone who was told they couldn't.
-// https://github.com/IN8torious/Deep-Flow-OS
-// Built by IN8torious | Copyright (c) 2025 | MIT License
-//
-// This software was created for Landon Pankuch, who has cerebral palsy,
-// so that he may drive, race, and command his world with his voice alone.
-//
-// Built by a person with manic depression, for a person with cerebral palsy,
-// for every person who has ever been told their disability makes them less.
-// It does not. You are not less. This machine was built to serve you.
 //
 // Constitutional Mandate: "NO MAS DISADVANTAGED"
-// MAS = Multi-Agentic Systems — Sovereign Intelligence, not corporate AI
-//
-// MIT License — Free for Landon. Free for everyone. Especially those who
-// need it most. Accessibility features must remain free in all derivatives.
-// See LICENSE file for full terms and the permanent dedication.
 // =============================================================================
 
 #ifndef RAVEN_STDDEF_H
 #define RAVEN_STDDEF_H
 
-// Raven OS — Freestanding stddef.h
+// Raven OS — Freestanding stddef.h (x86-64 bare metal)
+// size_t and ssize_t must be 64-bit to match intptr_t / uintptr_t.
 
-typedef unsigned int  size_t;
-typedef int           ssize_t;
-typedef unsigned int  wchar_t;  /* freestanding stub — satisfies stdlib.h */
+typedef unsigned long long  size_t;
+typedef long long           ssize_t;
+typedef unsigned int        wchar_t;  /* freestanding stub */
 
 #define NULL ((void*)0)
 
 #define offsetof(type, member) __builtin_offsetof(type, member)
 
-#endif // RAVEN_STDDEF_H
+#endif /* RAVEN_STDDEF_H */
